+++ Common/src/main/java/com/almostreliable/unified/compat/unification/CreateRecipeUnifier.java
@@ -1,6 +1,7 @@
 package com.almostreliable.unified.compat.unification;
 
 import com.almostreliable.unified.api.unification.bundled.GenericRecipeUnifier;
+import com.almostreliable.unified.api.unification.bundled.ShapedRecipeUnifier;
 import com.almostreliable.unified.api.unification.recipe.RecipeJson;
 import com.almostreliable.unified.api.unification.recipe.RecipeUnifier;
 import com.almostreliable.unified.api.unification.recipe.UnificationHelper;
@@ -15,6 +16,20 @@ public class CreateRecipeUnifier implements RecipeUnifier {
 
     @Override
     public void unify(UnificationHelper helper, RecipeJson recipe) {
+        // inherited crafting recipes
+        if (ShapedRecipeUnifier.isApplicable(recipe)) {
+            ShapedRecipeUnifier.INSTANCE.unify(helper, recipe);
+            return;
+        }
+
+        // sequenced assembly recipes
+        unifySequencedAssemblyRecipes(helper, recipe);
+
+        // all other recipes
+        GenericRecipeUnifier.INSTANCE.unify(helper, recipe);
+    }
+
+    private void unifySequencedAssemblyRecipes(UnificationHelper helper, RecipeJson recipe) {
         if (!(recipe.getProperty(SEQUENCE) instanceof JsonArray sequenceArray)) {
             return;
         }